:truck: move CertificationResult to certification
