+++ api/src/certification/results/domain/models/CertificationResult.js
@@ -1,8 +1,8 @@
 import _ from 'lodash';
 
-import { CompetenceMark } from '../../../certification/shared/domain/models/CompetenceMark.js';
-import { ComplementaryCertificationCourseResult } from '../../../certification/shared/domain/models/ComplementaryCertificationCourseResult.js';
-import { JuryComment, JuryCommentContexts } from '../../../certification/shared/domain/models/JuryComment.js';
+import { CompetenceMark } from '../../../shared/domain/models/CompetenceMark.js';
+import { ComplementaryCertificationCourseResult } from '../../../shared/domain/models/ComplementaryCertificationCourseResult.js';
+import { JuryComment, JuryCommentContexts } from '../../../shared/domain/models/JuryComment.js';
 
 /**
  * @readonly
