+++ api/src/certification/results/infrastructure/repositories/certification-result-repository.js
@@ -1,6 +1,6 @@
 import { knex } from '../../../../../db/knex-database-connection.js';
-import { CertificationResult } from '../../../../shared/domain/models/CertificationResult.js';
 import { ComplementaryCertificationCourseResult } from '../../../shared/domain/models/ComplementaryCertificationCourseResult.js';
+import { CertificationResult } from '../../domain/models/CertificationResult.js';
 
 const findBySessionId = async function ({ sessionId }) {
   const certificationResultDTOs = await _selectCertificationResults()
