+++ api/tests/certification/results/unit/domain/models/CertificationResult_test.js
@@ -1,6 +1,6 @@
-import { AutoJuryCommentKeys } from '../../../../src/certification/shared/domain/models/JuryComment.js';
-import { CertificationResult } from '../../../../src/shared/domain/models/CertificationResult.js';
-import { domainBuilder, expect } from '../../../test-helper.js';
+import { CertificationResult } from '../../../../../../src/certification/results/domain/models/CertificationResult.js';
+import { AutoJuryCommentKeys } from '../../../../../../src/certification/shared/domain/models/JuryComment.js';
+import { domainBuilder, expect } from '../../../../../test-helper.js';
 
 const CERTIFICATION_RESULT_STATUS_CANCELLED = CertificationResult.status.CANCELLED;
 const CERTIFICATION_RESULT_STATUS_ERROR = CertificationResult.status.ERROR;
