+++ api/tests/certification/results/unit/infrastructure/utils/csv/certification-results/CertificationResultsCsvValues_test.js
@@ -1,6 +1,6 @@
+import { CertificationResult } from '../../../../../../../../src/certification/results/domain/models/CertificationResult.js';
 import { CertificationResultsCsvValues } from '../../../../../../../../src/certification/results/infrastructure/utils/csv/certification-results/CertificationResultsCsvValues.js';
 import { AutoJuryCommentKeys } from '../../../../../../../../src/certification/shared/domain/models/JuryComment.js';
-import { CertificationResult } from '../../../../../../../../src/shared/domain/models/CertificationResult.js';
 import { getI18n } from '../../../../../../../../src/shared/infrastructure/i18n/i18n.js';
 import { domainBuilder, expect } from '../../../../../../../test-helper.js';
