+++ api/tests/tooling/domain-builder/factory/build-certification-result.js
@@ -1,4 +1,4 @@
-import { CertificationResult } from '../../../../src/shared/domain/models/CertificationResult.js';
+import { CertificationResult } from '../../../../src/certification/results/domain/models/CertificationResult.js';
 
 const buildCertificationResult = function ({
   id = 123,
